requirements.txt
Flask>=2.0
Werkzeug>=2.0
openpyxl
Pillow
pywebview
PyMuPDF
pandas
pytesseract
requests